+++ src/qwt_grid_raster_data.cpp
@@ -1,4 +1,8 @@
-﻿#include "qwt_grid_raster_data.h"
+﻿/******************************************************************************
+ * Qwt Widget Library
+ * Copyright (C) 2024   ChenZongYan <czy.t@163.com>
+ *****************************************************************************/
+#include "qwt_grid_raster_data.h"
 #include "qwt_interval.h"
 #include <QVector>
 #include <QRectF>